temp fix bug of clang compiling error: __compiletime_assert_* in i915 code
The root cause is the different behavior of the option '-fno-builtin' in
gcc and clang.

* In gcc, '-fno-builtin' prevents the compiler from transforming
  functions like 'memset' to '__builtin_memset', but explicit calls to
    '__builtin_memset' are still considered builtin functions with related
      optimizations enabled.

      * In clang, '-fno-builtin' seems to 'disable special handling ** and
        optimizations ** of builtin functions'.

The i915 code, on the other hand, relies on the builtin function
optimizations to eliminate the BUILD_BUGs. Since clang disables the
optimizations, these BUILD_BUGs still exist in the object files and
trigger link errors as intended.

This magic is introduced during 4.12-rc1 and seems to be the first and
currently only case throughout the kernel. Thus the difference of
'-fno-builtin' is not revealed till now.

Junjie Mao and I have not yet found any clang option that is exactly equivalent to
gcc '-fno-builtin'. A easy workaround is just to change macro BUILD_BUG() to BUG().