block/mq: do not invoke preempt_disable() AND don't complete requests via IPI
preempt_disable() and get_cpu() don't play well together with the sleeping
locks it tries to allocate later.
It seems to be enough to replace it with get_cpu_light() and migrate_disable().

The IPI runs in hardirq context and there are sleeping locks. This patch
moves the completion into a workqueue.